new mode for multiSANA
