+++ MPI/cmd_multi/random-align.sh
