+++ MPI/cmd_multi/run.sh
